requirements.txt
Flask
mysql-connector-python
scikit-learn
nltk
python-dotenv
google-genai
gunicorn